Adding Wagyu Key Gen option to the generating keys section
