+++ src/intl/compiled/en.json
@@ -407,6 +407,12 @@
       "value": "execution layer"
     }
   ],
+  "2Hk0nB": [
+    {
+      "type": 0,
+      "value": "Follow the instructions presented to you in the application to generate your keys."
+    }
+  ],
   "2PI6Di": [
     {
       "type": 0,
@@ -651,6 +657,12 @@
       "value": "Commitment"
     }
   ],
+  "5H/J84": [
+    {
+      "type": 0,
+      "value": "How do you want to generate your keys?"
+    }
+  ],
   "5HJLBU": [
     {
       "type": 0,
@@ -871,6 +883,12 @@
     {
       "type": 0,
       "value": "An Ethereum client designed to be enterprise-friendly for both public and private, permissioned network use cases."
+    }
+  ],
+  "8yom7x": [
+    {
+      "type": 0,
+      "value": "Run the Wagyu Key Gen app from withing Applications by right clicking and clicking Open. You will get a warning stating macOS cannot verify the developer of “Wagyu Key Gen.app”. Are you sure you want to open it?. Click Open and the app will open."
     }
   ],
   "9+Ddtu": [
@@ -1581,6 +1599,12 @@
       "value": "Generate keys"
     }
   ],
+  "GelBd9": [
+    {
+      "type": 0,
+      "value": "Make the file you just downloaded executable."
+    }
+  ],
   "GsBRWL": [
     {
       "type": 0,
@@ -1901,6 +1925,16 @@
     {
       "type": 0,
       "value": "Error"
+    }
+  ],
+  "KWWsDo": [
+    {
+      "type": 0,
+      "value": "Please make sure that you are downloading from the official StakeHouse GitHub account by verifying the url: "
+    },
+    {
+      "type": 1,
+      "value": "url"
     }
   ],
   "KjcAm2": [
@@ -2585,12 +2619,6 @@
       "value": "Verify it with the following command to check if it returns the client version correctly:"
     }
   ],
-  "S0e/cN": [
-    {
-      "type": 0,
-      "value": "How do you want to run the CLI?"
-    }
-  ],
   "S943uU": [
     {
       "type": 0,
@@ -2607,6 +2635,12 @@
     {
       "type": 0,
       "value": "Your stake has reached the deposit contract!"
+    }
+  ],
+  "SHNw06": [
+    {
+      "type": 0,
+      "value": "Execute the file you just downloaded."
     }
   ],
   "SR0qHv": [
@@ -3207,6 +3241,20 @@
       "value": " service."
     }
   ],
+  "YQJy4H": [
+    {
+      "type": 0,
+      "value": "You should now have your mnemonic written down in a safe place and a keystore saved for each of your "
+    },
+    {
+      "type": 1,
+      "value": "validatorCount"
+    },
+    {
+      "type": 0,
+      "value": " validators. Please make sure you keep these safe, preferably offline. Your validator keystores should be available in the selected directory."
+    }
+  ],
   "Yganfl": [
     {
       "type": 0,
@@ -3301,6 +3349,20 @@
     {
       "type": 0,
       "value": "Why is there a wait?"
+    }
+  ],
+  "ZlLjuq": [
+    {
+      "type": 0,
+      "value": "Please make sure you select "
+    },
+    {
+      "type": 1,
+      "value": "network"
+    },
+    {
+      "type": 0,
+      "value": " when prompted for a network, otherwise the deposit will be invalid."
     }
   ],
   "Zspg8H": [
@@ -3415,6 +3477,12 @@
       "value": "Besu installation documentation"
     }
   ],
+  "bUn2/Z": [
+    {
+      "type": 0,
+      "value": "Download Wagyu Key Gen app"
+    }
+  ],
   "bVtlF7": [
     {
       "type": 0,
@@ -3515,6 +3583,12 @@
     {
       "type": 0,
       "value": "Back"
+    }
+  ],
+  "d4g3yx": [
+    {
+      "type": 0,
+      "value": "Step 2: Generate deposit keys using the Wagyu Key Gen app"
     }
   ],
   "d5KEM8": [
@@ -4699,6 +4773,12 @@
       "value": "!"
     }
   ],
+  "pHQwsN": [
+    {
+      "type": 0,
+      "value": "Launch the app from your desktop environment by double clicking on it."
+    }
+  ],
   "pXyH++": [
     {
       "type": 0,
@@ -5001,6 +5081,12 @@
       "value": "."
     }
   ],
+  "sj/Zef": [
+    {
+      "type": 0,
+      "value": "Download Key Gen GUI app"
+    }
+  ],
   "smD5l/": [
     {
       "type": 0,
@@ -5275,6 +5361,12 @@
     {
       "type": 0,
       "value": "Proceed with caution. Our on-chain data source is down and we are unable to flag any double deposits."
+    }
+  ],
+  "vV6ua0": [
+    {
+      "type": 0,
+      "value": "Step 1: Download the Wagyu Key Gen app for your operating system"
     }
   ],
   "vY2wig": [
