+++ src/pages/GenerateKeys/Option1.tsx
@@ -6,19 +6,13 @@ import { Heading } from '../../components/Heading';
 import { Text } from '../../components/Text';
 import { Link } from '../../components/Link';
 import { Alert } from '../../components/Alert';
-import { Code } from '../../components/Code';
-import { NETWORK_NAME, IS_MAINNET } from '../../utils/envVars';
+import { NETWORK_NAME } from '../../utils/envVars';
 import { Button } from '../../components/Button';
-import githubScreenshot from '../../static/github-cli-screenshot.png';
-import { colors } from '../../styles/styledComponentsTheme';
+import githubScreenshot from '../../static/github-wagyu-key-gen-screenshot.png';
 
 const AlertIcon = styled(p => <GrommetAlert {...p} />)`
   display: block;
   margin: 1.3rem;
-`;
-
-const Pre = styled.pre`
-  white-space: normal;
 `;
 
 const GithubScreenshot = styled.img.attrs({ src: githubScreenshot })`
@@ -26,26 +20,20 @@ const GithubScreenshot = styled.img.attrs({ src: githubScreenshot })`
   width: 100%;
 `;
 
-export const Option1 = ({
-  validatorCount,
-  os,
-}: {
-  validatorCount: number | string;
-  os: string;
-}) => {
+export const Option1 = ({ os }: { os: string }) => {
   const { formatMessage } = useIntl();
 
   return (
     <div className="mt30">
       <Heading level={2} size="small" color="blueMedium" className="mb20">
-        <FormattedMessage defaultMessage="Download command line app" />
+        <FormattedMessage defaultMessage="Download Wagyu Key Gen app" />
       </Heading>
       <Text weight={500}>
-        <FormattedMessage defaultMessage="Step 1: Download the deposit command line interface app for your operating system" />
+        <FormattedMessage defaultMessage="Step 1: Download the Wagyu Key Gen app for your operating system" />
       </Text>
       <Link
         isTextLink={false}
-        to="https://github.com/ethereum/eth2.0-deposit-cli/releases/"
+        to="https://github.com/stake-house/wagyu-key-gen/releases"
         className="my40"
       >
         <Button
@@ -65,12 +53,12 @@ export const Option1 = ({
             style={{ wordBreak: 'break-word' }}
           >
             <FormattedMessage
-              defaultMessage="Please make sure that you are downloading from the official Ethereum
-              Foundation GitHub account by verifying the url: {url}"
+              defaultMessage="Please make sure that you are downloading from the official StakeHouse
+              GitHub account by verifying the url: {url}"
               values={{
                 url: (
                   <strong>
-                    https://github.com/ethereum/eth2.0-deposit-cli/releases/
+                    https://github.com/stake-house/wagyu-key-gen/releases
                   </strong>
                 ),
               }}
@@ -83,81 +71,51 @@ export const Option1 = ({
       <GithubScreenshot />
 
       <Text weight={500} className="mt20">
-        <FormattedMessage defaultMessage="Step 2: Generate deposit keys using the Ethereum Foundation deposit tool" />
+        <FormattedMessage defaultMessage="Step 2: Generate deposit keys using the Wagyu Key Gen app" />
       </Text>
       <Alert className="my20" variant="info">
         <FormattedMessage defaultMessage="For security, we recommend you disconnect from the internet to complete this step." />
       </Alert>
+
       <ul>
-        <li>
-          <FormattedMessage defaultMessage="Decompress the file you just downloaded" />
-        </li>
-        <li>
-          <FormattedMessage
-            defaultMessage="Use the terminal to move into the directory that contains the {deposit} executable"
-            values={{
-              deposit: <code>deposit</code>,
-            }}
-            description="{deposit} = 'deposit' styled as code"
-          />
-        </li>
-        <li>
-          <FormattedMessage defaultMessage="Run the following command to launch the app" />
-        </li>
-        <Alert variant="secondary" className="my10">
-          <Pre className="my10">
-            {(os === 'linux' || os === 'mac') && (
-              <span style={{ color: colors.red.medium }}>./deposit </span>
-            )}
-            {os === 'windows' && (
-              <>
-                <span style={{ color: colors.red.medium }}>.\deposit</span>
-                <span style={{ color: colors.purple.dark }}>.exe </span>
-              </>
-            )}
-            <span style={{ color: colors.red.medium }}>new-mnemonic</span>
-            <span style={{ color: colors.red.medium }}>
-              {validatorCount > 0
-                ? ` --${formatMessage({
-                    defaultMessage: 'num_validators',
-                    description:
-                      'this is used as a command line flag, short for "number of validators"',
-                  })} ${validatorCount}`
-                : ''}{' '}
-            </span>
-            <span style={{ color: colors.red.medium }}>
-              {`--${formatMessage({
-                defaultMessage: 'chain',
-                description: 'this is used as a command line flag',
-              })} ${NETWORK_NAME.toLowerCase()}`}
-            </span>
-          </Pre>
-        </Alert>
+        {os === 'windows' && (
+          <li>
+            <FormattedMessage defaultMessage="Execute the file you just downloaded." />
+          </li>
+        )}
+        {os === 'linux' && (
+          <>
+            <li>
+              <FormattedMessage defaultMessage="Make the file you just downloaded executable." />
+            </li>
+            <li>
+              <FormattedMessage defaultMessage="Launch the app from your desktop environment by double clicking on it." />
+            </li>
+          </>
+        )}
+        {os === 'mac' && (
+          <>
+            <li>
+              <FormattedMessage defaultMessage="Execute the file you just downloaded." />
+            </li>
+            <li>
+              <FormattedMessage defaultMessage="Run the Wagyu Key Gen app from withing Applications by right clicking and clicking Open. You will get a warning stating macOS cannot verify the developer of “Wagyu Key Gen.app”. Are you sure you want to open it?. Click Open and the app will open." />
+            </li>
+          </>
+        )}
         <Alert variant="error" className="my10">
           <Text>
             <FormattedMessage
-              defaultMessage="Please make sure you have set {flag} for {network}, otherwise the deposit will be invalid."
+              defaultMessage="Please make sure you select {network} when prompted for a network, otherwise the deposit will be invalid."
               values={{
-                flag: (
-                  <Code>
-                    {`--${formatMessage({
-                      defaultMessage: 'chain',
-                      description: 'this is used as a command line flag',
-                    })} ${NETWORK_NAME.toLowerCase()}`}
-                  </Code>
-                ),
-                network: (
-                  <span>
-                    {IS_MAINNET ? NETWORK_NAME : `${NETWORK_NAME} testnet`}
-                  </span>
-                ),
+                network: <span>{NETWORK_NAME}</span>,
               }}
               description="{flag} and {network} are terminal commands styled as code."
             />
           </Text>
         </Alert>
         <li>
-          <FormattedMessage defaultMessage="Now follow the instructions presented to you in the terminal window to generate your keys." />
+          <FormattedMessage defaultMessage="Follow the instructions presented to you in the application to generate your keys." />
         </li>
       </ul>
     </div>
